Improved ProtoBuf serialization
The serialization of 'environment' and 'language' objects changed
considerably between RProtoBuf package versions 0.4.1 and 0.4.2. In
brief, the former skips them with a warning message, whereas the
latter serializes them as native byte arrays(?). Unfortunately, the
JPMML-Converter library cannot handle this encoding.
The workaround is to convert all problematic objects to character
vectors using the 'format' function.